[refactor] Frame의 frameNumber 제거
