+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,13 +1,11 @@
 package bowling.domain.frame;
 
 import bowling.annotations.GetterForUI;
-import bowling.domain.value.FrameNumber;
 import bowling.domain.value.Pins;
 import bowling.domain.value.Score;
 
 public abstract class Frame {
     protected Score score;
-    protected FrameNumber frameNumber;
 
     public abstract boolean isFrameOver();
 
@@ -21,11 +19,6 @@ public abstract class Frame {
 
     public void accumulateScore(int countOfPins) {
         this.score = score.accumulateScore(countOfPins);
-    }
-
-    @GetterForUI
-    public final FrameNumber getCurrentFrameNumber() {
-        return frameNumber;
     }
 
     @GetterForUI
