+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -9,14 +9,13 @@ import bowling.domain.value.Score;
 public class NormalFrame extends Frame {
     private State state;
 
-    private NormalFrame(FrameNumber frameNumber) {
+    private NormalFrame() {
         this.score = Score.init();
         this.state = Ready.of();
-        this.frameNumber = frameNumber;
     }
 
-    public static Frame create(FrameNumber frameNumber) {
-        return new NormalFrame(frameNumber);
+    public static Frame create() {
+        return new NormalFrame();
     }
 
     @Override
